+++ apps/app/src/__tests__/account-queries-source.test.ts
@@ -11,7 +11,6 @@ function source(path: string) {
 const migratedFiles = [
   "src/components/app-sidebar.tsx",
   "src/account/settings/profile-data-display.tsx",
-  "src/account/settings/account-settings-actions.ts",
   "src/account/tasks/username-account-task-client.tsx",
 ] as const;
